refactor(#121): Phase2 Step2.3 — move archive_usage core to analyzer/archive/usage.py
scripts/archive_usage.py is now a thin dispatch shim; main()/argparse and
all core logic live in analyzer/archive/usage.py. Phase 2 complete.
